requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
packaging==23.2
python-dotenv==1.0.1
SQLAlchemy==2.0.27
typing_extensions==4.9.0
Werkzeug==3.0.1
yarl==1.9.4
